refactor(terminal-workspace): extract renderer read models
