+++ test/renderer/features/terminal-workspace/TerminalWorkspacePanelInternals.fixture-e2e.test.ts
@@ -1,4 +1,5 @@
 import { normalizeTerminalCommandRunEventDetail } from '@features/terminal-workspace/renderer/adapters/terminalCommandRunEvents';
+import { resolveTerminalLocalAutocompleteSuggestion } from '@features/terminal-workspace/renderer/model/terminalCommandAutocomplete';
 import {
   closeSupersededTerminalCommandRuns,
   inferTerminalCommandCompletion,
@@ -10,8 +11,7 @@ import {
 import {
   formatTerminalPromptLabel,
   formatWorkingDirectory,
-  resolveTerminalLocalAutocompleteSuggestion,
-} from '@features/terminal-workspace/renderer/ui/TerminalWorkspacePanel';
+} from '@features/terminal-workspace/renderer/model/terminalPathPresentation';
 import { describe, expect, it } from 'vitest';
 
 describe('terminal workspace panel internals fixture-e2e', () => {
